fixed removing workinghour implementation and api, added getIssues api

+++ src/main/java/rest/WorkingHourApi.java
@@ -11,6 +11,7 @@ import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.core.Context;
@@ -20,6 +21,7 @@ import javax.ws.rs.core.Response;
 import authentication.Secured;
 import authentication.UserSecurityContext;
 import entities.DebugLogger;
+import entities.Issue;
 import entities.User;
 import entities.WorkingHour;
 import service.WorkingHourService;
@@ -60,15 +62,19 @@ public class WorkingHourApi {
 	@Secured
 	@DELETE
 	@Consumes(MediaType.APPLICATION_JSON)
-	@Produces(MediaType.APPLICATION_JSON)
-	public Response removeWorkingHour(@Context ContainerRequestContext securityContext, WorkingHour workingHour) {
-		UserSecurityContext userSecurityContext = (UserSecurityContext) securityContext.getSecurityContext();
-		User user = userSecurityContext.getUser();
+	@Produces(MediaType.TEXT_PLAIN)
+	@Path("/{id}")
+	public Response removeWorkingHour(@PathParam(value = "id") long id) {
+		WorkingHour workingHour = hourService.getWorkingHourById(Long.valueOf(id));
+		if (workingHour == null) {
+			logger.log("Not removing from database, because WorkingHour not exists with id: " + id);
+			return Response.notModified().build();
+		}
+		logger.log("Removing from database: " + workingHour);
 		hourService.removeWorkingHour(workingHour);
-		return Response.ok("Working hour removed to user: " + user.getName() + "with start: " + workingHour.getStarting()
-		+ " with duration: " + workingHour.getDuration()).build();
+		return Response.ok().build();
 	}
-	
+
 	@GET
 	@Secured
 	@Produces(MediaType.APPLICATION_JSON)
@@ -80,5 +86,20 @@ public class WorkingHourApi {
 			workingHours = new ArrayList<>();
 		}
 		return Response.ok(workingHours).build();
+	}
+
+	@GET
+	@Secured
+	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/issues")
+	public Response getIssues() {
+		List<Issue> issues = hourService.getIssues();
+		if (issues == null) {
+			issues = new ArrayList<>();
+			logger.log("No issues found. Returning empty list.");
+		} else {
+			logger.log("Issues found, returning count: " + issues.size());
+		}
+		return Response.ok(issues).build();
 	}
 }
